Fix BPM to be a double

+++ api/Features/Tracks/TrackRepository.cs
@@ -68,7 +68,7 @@ public sealed class TrackRepository(IConfiguration configuration) : BaseReposito
         artistField.AnalyzerName = LexicalAnalyzerName.StandardAsciiFoldingLucene;
 
         var albumField = new SimpleField(nameof(Track.Album), SearchFieldDataType.String);
-        var bpmField = new SimpleField(nameof(Track.BPM), SearchFieldDataType.Int32) { IsFilterable = true };
+        var bpmField = new SimpleField(nameof(Track.BPM), SearchFieldDataType.Double) { IsFilterable = true };
         var timeField = new SimpleField(nameof(Track.Time), SearchFieldDataType.String);
         var keyField = new SimpleField(nameof(Track.Key), SearchFieldDataType.String);
